Add hit counters leaderboard and instructions
The instructions say to use the exact name and capitalization for
the URL, but it's not actually important for the capitalization to
be the same. I just want people to think it's important to be exact
so they'll enter the name instead of a nickname so I can link to
the page from the leaderboard.